requirements.txt
reflex==0.8.26
hubspot-api-client
openpyxl